Thu, Dec 15, 2022, 4:00 PM +00:00

+++ fuel/fuel.py
@@ -6,7 +6,7 @@ try:
         print ("F")
     elif (int(Fraction2[0])/int(Fraction2[1]))*100 <= int("1"):
         print ("E")
-    else: print((float(Fraction2[0]) / float(Fraction2[1])) * 100 )
+    else: print(int(float(Fraction2[0]) / float(Fraction2[1])) * 100 )
 
 
 except :
